Moved DCM-specific function from can_actions to DCM module
